requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
pymysql==1.1.0

# Password hashing
bcrypt==4.1.2

# JWT Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# HTTP client (for testing and external API calls)
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0 